senado-graph/scraper/requirements.txt
beautifulsoup4>=4.12.0
requests>=2.31.0
neo4j>=5.14.0
python-dotenv>=1.0.0
lxml>=4.9.0
pytest>=7.4.0
pytest-asyncio>=0.21.0